implement bitmap allocation layer
- bitmap creation/loading/flushing
- bitmap block allocation
- bit manipulation helpers
- test

+++ include/superblock.h
@@ -9,7 +9,7 @@
 struct superblock {
     uint64_t identifier;
     uint64_t block_size;
-    uint64_t block_count;
+    size_t block_count;
     uint64_t root_dir_index;
     uint64_t bitmap_index;
 };
